requirements.txt
os
sty
discord
discord.ext
datetime
json
pymongo
googleapiclient.discovery
dotenv
asyncio
threading
time